# docs/openapi/Everdue.Server.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Everdue.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/settings/channels/{channel}": {
      "put": {
        "tags": [
          "Channels"
        ],
        "summary": "Saves this tenant's credentials for one channel.",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelSettingsDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Channels"
        ],
        "summary": "Removes this tenant's credentials. Any system-scope fallback applies again afterwards.",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/settings/channels/{channel}/test": {
      "post": {
        "tags": [
          "Channels"
        ],
        "summary": "Sends one message to the caller — the only honest way to test a channel.",
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelTestResultDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelSettingsDto": {
        "required": [
          "channel",
          "configured",
          "active",
          "usingSystemScope",
          "summary",
          "redactedConfigJson",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/NotificationChannel"
          },
          "configured": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          },
          "usingSystemScope": {
            "type": "boolean"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "redactedConfigJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "updatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "ChannelTestResultDto": {
        "required": [
          "sent",
          "error"
        ],
        "type": "object",
        "properties": {
          "sent": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "NotificationChannel": {
        "enum": [
          "Email",
          "Telegram",
          "WhatsApp"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveChannelRequest": {
        "required": [
          "configJson",
          "active"
        ],
        "type": "object",
        "properties": {
          "configJson": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}